requirements.txt
openai>=1.35.0
requests>=2.32.0
python-dotenv>=1.0.1
